requirements.txt
presidio_analyzer
opencv-python
pytesseract
click
ffmpeg-python